add modifications based on tasks.

+++ app/Http/Controllers/User/DeleteAction.php
@@ -6,8 +6,10 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Repository\UserRepository;
 use DomainException;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use InvalidArgumentException;
 
 class DeleteAction extends Controller
 {
@@ -16,14 +18,25 @@ class DeleteAction extends Controller
     )
     {}
 
-    public function __invoke()
+    public function __invoke(Request $request): JsonResponse
     {
-        $user = Auth::user();
-        if (!$user instanceof User) {
-            throw new DomainException('User is not a valid user');
+        if (!$request->has('user_id')){
+            throw new InvalidArgumentException('Not all required parameters are in the request', 404);
         }
+
+        $userId = $request->get('user_id');
+
+        $user = $this->userRepository->find($userId);
+
+        if (!$user instanceof User) {
+            throw new DomainException('User is not a valid user', 404);
+        }
+
         $this->userRepository->delete($user);
-        return redirect('/');
+
+        return new JsonResponse([
+            'message' => 'User succesfully deleted',
+        ]);
     }
